Switch from Stripe Checkout to Elements
This would make payments happen on the same page as the creation of the
site. This should reduce friction and the number of redirects, thus the
time used to create a site. However, this may reduce the time too much
and not let the schema migrations be finished on time, so a warning is
added.

This new page uses JS for the form handling and payments. Obsolete pages
are deleted, and a new page is created for site creation with an
attached invoice when needed, bypassing the payment validation.