Return 204 for Stripe and use consumer to migrate
Stripe doesn't like the 404 status when the client does not exist, so
allow the webhook with an Empty message (204) which is acceptable.

Also, to avoid 503s, use the migration consumer instead of the two-step
approach when creating sites in the admin. As user is nullable, check in
the consumer. Also some quick touch-ups to the admin list.