Order debates in private URL ballots column by seq
The superseding debates queryset when adding the ballots link column to
the table is not sorted, unlike the squashed debates list. This sets the
ordering to be like its invocation, by round sequence.

To populate the debates with the existence of a ballot, an inplace
function is created like for populating them with confirmed ballots. A
"dummy" BallotSubmission is also created to find the class that the
ballot would have.

Fixes #1900